Refactor elements into Element.java, and merge the interfaces.
