--- src/main/java/powder/particles/IParticleInit.java
@@ -1,5 +1,0 @@
-package powder.particles;
-
-public interface IParticleInit {
-    void init(Particle p);
-}
