--- src/main/java/powder/particles/IParticleUpdate.java
@@ -1,5 +1,0 @@
-package powder.particles;
-
-public interface IParticleUpdate {
-    void update(Particle p);
-}
