+++ src/main/java/powder/particles/ParticleBehaviour.java
@@ -1,0 +1,12 @@
+package powder.particles;
+
+public abstract class ParticleBehaviour {
+    /**
+     * Initialises the particle - called when the particle is placed.
+     *
+     * @param p
+     */
+    public abstract void init(Particle p);
+
+    public abstract void update(Particle p);
+}
